algorithms/selfie2anime/requirements.txt

algorithmia>=1.0.0,<2.0
six
opencv-python-headless>=4.2,<4.3
tensorflow-gpu==1.14.0

